refactor: rename upload-extension to async-extension
