+++ packages/remirror__core/src/builtins/builtin-preset.ts
@@ -14,7 +14,7 @@ import { PluginsExtension } from './plugins-extension';
 import { SchemaExtension } from './schema-extension';
 import { SuggestExtension, SuggestOptions } from './suggest-extension';
 import { TagsExtension } from './tags-extension';
-import { UploadExtension } from './upload-extension';
+import { UploadExtension } from './async-extension';
 
 export interface BuiltinOptions
   extends SuggestOptions,
